mnist/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
matplotlib
python-dotenv>=0.5.1
torch
torchvision
numpy
Pillow
